# ENH: Optimized spatial join