--- a/geopandas/tools/sjoin.py
+++ b/geopandas/tools/sjoin.py
@@ -44,8 +44,6 @@ def sjoin(left_df, right_df, how='left', op='intersects', convert_crs=True, lsuf
                 right_df = right_df.to_crs(left_df.crs)
             elif left_df.values.nbytes < right_df.values.nbytes:
                 left_df = left_df.to_crs(right_df.crs)
-	else:
-	    pass
 
     # CONSTRUCT SPATIAL INDEX FOR RIGHT DATAFRAME
     tree_idx = rtree.index.Index()