@perrygeo 2015-04-08
Nice work @mdbartos ! The original work done at the SciPy sprint in 2014 was a good proof of concept but these performance gains look really significant and make geopandas a real contender for this type of analysis. 

I have some concerns with the `crs_convert` defaulting to true. It sets a precedent that every function which compares geometries should implement it's own crs conversion logic. I'd prefer if the `sjoin` function (and all functions for that matter) follow the single responsibility principle - `sjoin` should only perform a spatial join and any crs conversion of GeoDataFrames should be dealt with separately and explicitly. Beside the single responsibility principle, there are performance and potential inaccuracies that can arise when dealing with edge cases that make implicit, automatic reprojection a non-starter IMO. I'd prefer to, at most, use a warning to indicate a crs mismatch.