@mdbartos 2015-04-10
Greetings! Thanks for the positive feedback. After looking through your comments, I've incorporated a couple changes:

+ I've removed the `convert_crs` functionality. If there is a crs mismatch, a warning will be printed to the console but no conversion will take place.

+ I've set the default join logic to 'inner' (this is the default used by pandas' `merge` method, which is the pandas function that most resembles `sjoin`).

+ I've made efforts to maintain compliance with PEP8 and geopandas style guidelines.

Let me know if you have any other suggestions!

P.S. I'm wondering if there's a good way to run speed tests in travis? I want to be able to validate the speed increases that I'm seeing.

P.P.S. Note for future development: this code can probably be optimized further by figuring out a faster way to apply the predicate operations (e.g. 'contains', 'intersects', etc.) to the geometry arrays. Right now I use `np.vectorize()`, which works for any arbitrary function and any arbitrary datatype, but is essentially a "for" loop under the hood. Personally, I don't know if there is a faster way to apply arbitrary functions to an ndarray containing 'object'-type data, but it's something worth investigating.