@kjordahl 2015-04-10
@mdbartos This looks great! I will try to give this a proper review in the next few days.

To your question about running speed tests on Travis, although it is possible, it isn't really reliable, as you don't have any guarantees about the hardware it is running on, other processes on the machine, etc. I would love to try out [Airspeed Velocity](https://github.com/spacetelescope/asv/) for benchmarking over time, I think that would be well suited to this question.