@kjordahl 2015-04-13
:+1: This is really nice. My trivial comments above just regard comments, and can be cleaned up later.

The warning behavior for mismactched crs is consistent with what we do for other geo operations. We may think about enforcing it more strongly in the future.